Pull out all unit-related code into a new 'units' package.
