+++ server/units_resource.go
@@ -4,6 +4,8 @@ import (
 	"log"
 	"net/http"
 	"net/url"
+
+	"github.com/bmorton/deployster/units"
 )
 
 // UnitsResource is the HTTP resource responsible for getting basic information
@@ -15,7 +17,7 @@ type UnitsResource struct {
 // UnitsResponse is the wrapper struct for the JSON payload returned by the
 // Index action.
 type UnitsResponse struct {
-	Units []VersionedUnit `json:"units"`
+	Units []units.VersionedUnit `json:"units"`
 }
 
 // Index is the GET endpoint for listing all units that exist for a given
@@ -28,12 +30,12 @@ func (ur *UnitsResource) Index(u *url.URL, h http.Header, req interface{}) (int,
 	statusCode := http.StatusOK
 	response := &UnitsResponse{}
 
-	units, err := ur.Fleet.Units()
+	allUnits, err := ur.Fleet.Units()
 	if err != nil {
 		log.Println(err)
 		return http.StatusInternalServerError, nil, nil, err
 	}
-	response.Units = FindServiceUnits(u.Query().Get("name"), "", units)
+	response.Units = units.FindServiceUnits(u.Query().Get("name"), "", allUnits)
 
 	return statusCode, nil, response, nil
 }
