+++ units/versioned_unit.go
@@ -1,4 +1,4 @@
-package server
+package units
 
 import (
 	"fmt"
