+++ units/versioned_unit_test.go
@@ -1,4 +1,4 @@
-package server
+package units
 
 import (
 	"testing"
